커뮤니티 총 탄소절약량 로직/헤더부분 수정

+++ creen_carry_back/src/main/java/kr/co/iei/member/model/dao/MemberDao.java
@@ -46,7 +46,7 @@ public interface MemberDao {
 
 	Double getTotalCarbonPoint(String memberId);
 
-	int getCommunityTotalCarbon();
+	Double getCommunityTotalCarbon();
 
 	int checkActiveOrder(String memberId);
