+++ creen_carry_back/src/main/java/kr/co/iei/member/model/service/MemberService.java
@@ -227,7 +227,7 @@ public class MemberService {
         return memberDao.getTotalCarbonPoint(memberId);
     }
 
-    public int getCommunityTotalCarbon() {
+    public Double getCommunityTotalCarbon() {
         return memberDao.getCommunityTotalCarbon();
     }
